Add test to check sqlite3 sends service_info
