+++ tests/contrib/sqlite3/test_sqlite3.py
@@ -6,6 +6,7 @@ import time
 from nose.tools import eq_
 
 # project
+import ddtrace
 from ddtrace import Pin
 from ddtrace.contrib.sqlite3 import connection_factory
 from ddtrace.contrib.sqlite3.patch import patch, unpatch
@@ -30,6 +31,22 @@ class TestSQLite(object):
 
     def tearDown(self):
         unpatch()
+
+    def test_service_info(self):
+        tracer = get_dummy_tracer()
+        backup_tracer = ddtrace.tracer
+        ddtrace.tracer = tracer
+
+        db = sqlite3.connect(":memory:")
+
+        services = tracer.writer.pop_services()
+        eq_(len(services), 1)
+        expected = {
+            'sqlite': {'app': 'sqlite', 'app_type': 'db'}
+        }
+        eq_(expected, services)
+
+        ddtrace.tracer = backup_tracer
 
     def test_sqlite(self):
         tracer = get_dummy_tracer()
